Menambahkan image dasboard dan history ketika data kosong
