+++ app/src/main/java/com/example/myapplication/ui/main/signup/SignUpPresenter.kt
@@ -1,0 +1,6 @@
+package com.example.myapplication.ui.main.signup
+
+interface SignUpPresenter {
+
+    fun signUp(username: String, password: String, email: String)
+}
