+++ app/src/main/java/com/example/myapplication/ui/main/signup/SignUpView.kt
@@ -1,0 +1,9 @@
+package com.example.myapplication.ui.main.signup
+
+interface SignUpView {
+
+    fun onSuccess()
+
+    fun onError(msg: String)
+
+}
