--- app/src/main/java/com/example/myapplication/ui/signup/SignUpPresenter.kt
@@ -1,6 +1,0 @@
-package com.example.myapplication.ui.signup
-
-interface SignUpPresenter {
-
-    fun signUp(username: String, password: String, email: String)
-}
