--- app/src/main/java/com/example/myapplication/ui/signup/SignUpView.kt
@@ -1,9 +1,0 @@
-package com.example.myapplication.ui.signup
-
-interface SignUpView {
-
-    fun onSuccess()
-
-    fun onError(msg: String)
-
-}
